move these subscriptions into playlist validations
because really they're validations based on a response from the iframe
